+++ test/browser/svg.js
@@ -70,20 +70,20 @@ describe('svg', () => {
 				<path class={c && ('bar_'+c)} stroke="white" fill="black" d="M347.1 357.9L183.3 256.5 13 357.9V1.7h334.1v356.2zM58.5 47.2v231.4l124.8-74.1 118.3 72.8V47.2H58.5z" />
 			</svg>
 		);
-		let root = render(<Demo c="1" />, scratch, root);
+		let root = render(<Demo c="1" />, scratch, {merge:root});
 		sinon.spy(root, 'removeAttribute');
-		root = render(<Demo />, scratch, root);
+		root = render(<Demo />, scratch, {merge:root});
 		expect(root.removeAttribute).to.have.been.calledOnce.and.calledWith('class');
 		root.removeAttribute.restore();
 
-		root = render(<div />, scratch, root);
-		root = render(<Demo />, scratch, root);
+		root = render(<div />, scratch, {merge:root});
+		root = render(<Demo />, scratch,  {merge:root});
 		sinon.spy(root, 'setAttribute');
-		root = render(<Demo c="2" />, scratch, root);
+		root = render(<Demo c="2" />, scratch,  {merge:root});
 		expect(root.setAttribute).to.have.been.calledOnce.and.calledWith('class', 'foo_2');
 		root.setAttribute.restore();
-		root = render(<Demo c="3" />, scratch, root);
-		root = render(<Demo />, scratch, root);
+		root = render(<Demo c="3" />, scratch,  {merge:root});
+		root = render(<Demo />, scratch,  {merge:root});
 	});
 
 	it('should still support class attribute', () => {